Implement Load Wait / Remove Old Headings Check
