+++ src/utils.ts
@@ -43,14 +43,17 @@ function findLastFromindex<T = unknown>(
   return index;
 }
 
-export function calcIndentLevels(headings: { heading: HeadingCache; offset: number; }[]): number[] {
+export function calcIndentLevels(
+  headings: { heading: HeadingCache; offset: number }[]
+): number[] {
   const result: number[] = [];
   if (!headings.length) {
     return result;
   }
   const topLevelIndex = headings.reduce<number>(
-    (res, cur, i) => (cur.heading.level < headings[res].heading.level ? i : res),
-    0,
+    (res, cur, i) =>
+      cur.heading.level < headings[res].heading.level ? i : res,
+    0
   );
   result.push(
     ...calcIndentLevels(headings.slice(0, topLevelIndex)).map(
@@ -62,7 +65,11 @@ export function calcIndentLevels(headings: { heading: HeadingCache; offset: numb
       result.push(0);
       return;
     }
-    const parentIndex = findLastFromindex(list, i - 1, item => item.heading.level < heading.heading.level);
+    const parentIndex = findLastFromindex(
+      list,
+      i - 1,
+      (item) => item.heading.level < heading.heading.level
+    );
     if (parentIndex === -1) {
       result.push(0);
     } else {
@@ -70,4 +77,41 @@ export function calcIndentLevels(headings: { heading: HeadingCache; offset: numb
     }
   });
   return result;
+}
+
+export function trivial(
+  subHeadings: { heading: HeadingCache; offset: number }[],
+  result: { heading: HeadingCache; offset: number }[],
+  mode: 'default' | 'concise'
+) {
+  if (!subHeadings.length) {
+    return result;
+  }
+  const topLevel = subHeadings.reduce(
+    (res, cur) => Math.min(res, cur.heading.level),
+    6
+  );
+  const indexesOfTopLevel = subHeadings.reduce<number[]>(
+    (indexes, cur, index) => {
+      if (cur.heading.level === topLevel) {
+        indexes.push(index);
+      }
+      return indexes;
+    },
+    []
+  );
+  if (mode === 'concise') {
+    if (indexesOfTopLevel.length >= 1) {
+      result.push(subHeadings[indexesOfTopLevel[indexesOfTopLevel.length - 1]]);
+    }
+  } else {
+    for (const index of indexesOfTopLevel) {
+      result.push(subHeadings[index]);
+    }
+  }
+  trivial(
+    subHeadings.slice(indexesOfTopLevel[indexesOfTopLevel.length - 1] + 1),
+    result,
+    mode
+  );
 }
